refactor: remove obsolete helpers and stale tests
