+++ src/Foundry.Core/Services/Configuration/ConfigurationNavigationTarget.cs
@@ -1,0 +1,24 @@
+// Copyright (c) Foundry Project contributors.
+// Licensed under the MIT License.
+// See the LICENSE file in the project root for more information.
+
+namespace Foundry.Core.Services.Configuration;
+
+public enum ConfigurationNavigationTarget
+{
+    None,
+    Adk,
+    General,
+    EthernetDot1x,
+    Wifi,
+    AutopilotJsonProfile,
+    AutopilotHardwareHashUpload,
+    AutopilotInteractiveHardwareHashUpload,
+    OperatingSystemSelection,
+    Unattend,
+    MachineNaming,
+    Oobe,
+    WindowsOptionalFeatures,
+    AppxRemoval,
+    AiComponentRemoval
+}
